chore: Minor API modifications, lint checks for server code.
Also, noticed a bug in client code and addressed it here.  The main collection component was passing an undefiend collection id into the subject tag selector component.
