+++ package.json
@@ -43,17 +43,17 @@
     "babel-preset-es2015": "^6.1.18",
     "babel-preset-stage-0": "^6.1.18",
     "babel-register": "^6.7.2",
-    "browser-sync": "^2.5.2",
+    "browser-sync": "^2.18.6",
     "chai": "^3.5.0",
-    "concurrently": "^2.0.0",
+    "concurrently": "^3.1.0",
     "coveralls": "^2.11.2",
     "del": "^2.2.1",
-    "gulp": "^3.9.0",
+    "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
     "gulp-clean-css": "^2.0.3",
     "gulp-concat": "^2.4.3",
     "gulp-coveralls": "^0.1.3",
-    "gulp-htmlmin": "^2.0.0",
+    "gulp-htmlmin": "^3.0.0",
     "gulp-imagemin": "^3.1.1",
     "gulp-jshint": "^2.0.4",
     "gulp-mocha": "^3.0.1",
@@ -61,8 +61,8 @@
     "gulp-rename": "^1.2.0",
     "gulp-rev-append": "^0.1.6",
     "gulp-typescript": "~3.1.2",
-    "gulp-uglify": "^1.1.0",
-    "gulp-usemin": "^0.3.26",
+    "gulp-uglify": "^2.0.0",
+    "gulp-usemin": "^0.3.27",
     "gulp-util": "^3.0.3",
     "istanbul": "^0.4.4",
     "jasmine-core": "^2.2.0",
@@ -78,7 +78,7 @@
     "karma-opera-launcher": "^1.0.0",
     "karma-phantomjs-launcher": "^1.0.1",
     "karma-safari-launcher": "^1.0.0",
-    "mocha": "^2.4.5",
+    "mocha": "^3.2.0",
     "ng-annotate": "^1.2.1",
     "protractor": "^3.1.0",
     "require-dir": "^0.3.0",
