+++ server/api/tagger/controllers/collection.js
@@ -15,7 +15,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 /**
- * Controllers for the Tagger administrative API and for external clients.
+ * Controllers for the Tagger administrative API and external clients.
  *
  * @author Michael Spalti
  */
@@ -66,10 +66,10 @@ exports.setPublicationStatus = function (req, res) {
 
   taggerDao.setPublicationStatus(pubStatus, collectionId).then(
     function() {
-      utils.sendResponse(res, {status: 'success'})
+      utils.sendSuccessJson(res);
     }).error(function(err) {
-      console.log(err)
-  })
+      console.log(err);
+  });
 
 };
 
@@ -78,10 +78,10 @@ exports.getPublicationStatus = function (req, res) {
 
   taggerDao.getPublicationStatus(collectionId).then(
     function(collection) {
-      utils.sendResponse(res, collection)
+      utils.sendResponse(res, collection);
     }).error(function(err) {
-    console.log(err)
-  })
+    console.log(err);
+  });
 
 };
 
@@ -116,23 +116,6 @@ exports.list = function (req, res) {
   }).error(function (err) {
     console.log(err);
   });
-};
-
-/**
- * Retrieves areas by collection id for the administrative
- * collections panel.
- * @param req
- * @param res
- */
-exports.areas = function (req, res) {
-  const collId = req.params.collId;
-
-  taggerDao.findAreasForCollection(collId).then(function (areas) {
-    utils.sendResponse(res, areas);
-  }).error(function (err) {
-    console.log(err);
-  });
-
 };
 
 /**
@@ -185,7 +168,7 @@ exports.removeTypeTarget = function (req, res) {
   const typeId = req.params.typeId;
 
   taggerDao.deleteItemContentTarget(collId, typeId).then(function () {
-    utils.sendResponse(res, {status: 'success'});
+    utils.sendSuccessJson(res);
   }).error(function (e) {
     console.log(e);
   });
@@ -247,12 +230,41 @@ exports.removeTagTarget = function (req, res) {
   const tagId = req.params.tagId;
 
   taggerDao.deleteTagTarget(collId, tagId).then(function () {
-    utils.sendResponse(res, {status: 'success'});
+    utils.sendSuccessJson(res);
   }).error(function (e) {
     console.log(e);
   });
 
 };
+
+/**
+ * Retrieves areas by collection id for the administrative
+ * collections panel.
+ * @param req
+ * @param res
+ */
+exports.areas = function (req, res) {
+  const collId = req.params.collId;
+
+  taggerDao.findAreasForCollection(collId).then(function (areas) {
+    utils.sendResponse(res, areas);
+  }).error(function (err) {
+    console.log(err);
+  });
+
+};
+
+
+
+function _areaIdsForCollection (collId, callback) {
+
+  taggerDao.getAreaIdsForCollection(collId).then(function (result) {
+    callback(null, result);
+  })
+    .error(function (err) {
+      console.log(err);
+    });
+}
 
 /**
  * Adds a collection to a collection area.
@@ -275,22 +287,12 @@ function _addArea(collId, areaId, res) {
 
       },
       areaList: function (callback) {
-        taggerDao.getAreaIdsForCollection(collId)
-          .then(function (result) {
-            callback(null, result);
-          })
-          .error(function (err) {
-            console.log(err);
-          });
+        _areaIdsForCollection(collId, callback);
       }
     },
 
     function (err, result) {
-      if (err) {
-        console.log(err);
-      }
-      utils.sendResponse(res, {status: 'success', areaTargets: result.areaList});
-
+      utils.sendAreaTargetsReponse(res, err, result);
     }
   );
 }
@@ -308,17 +310,17 @@ exports.getFirstCollectionInArea = function (req, res) {
       taggerDao.findCollectionsInArea(areaId)
         .then(
           function(collections) {
-
             callback(null, collections[0].dataValues.CollectionId);
+
           }
-        )
+        );
     },
     function (collId, callback) {
       taggerDao.findCollectionById(collId)
         .then(
           function(collection) {
-
             callback(null, collection);
+
           }
         );
     }
@@ -399,20 +401,12 @@ exports.removeAreaTarget = function (req, res) {
           });
       },
       areaList: function (callback) {
-        taggerDao.getAreaIdsForCollection(collId).then(function (result) {
-          callback(null, result);
-        })
-          .error(function (err) {
-            console.log(err);
-          });
+        _areaIdsForCollection(collId, callback);
       }
     },
 
     function (err, result) {
-      if (err) {
-        console.log(err);
-      }
-      utils.sendResponse(res, {status: 'success', areaTargets: result.areaList});
+      utils.sendAreaTargetsReponse(res, err, result);
 
     });
 };
@@ -531,15 +525,16 @@ exports.update = function (req, res) {
     },
     function (err, result) {
       if (err !== undefined && err !== null) {
-        utils.sendResponse(res, {status: 'failed'});
         console.log(err);
+        utils.sendErrorJson(res, err);
+
       }
       // If no category exists for this collection,
       // add new entry.
       if (result.checkCategory === null) {
         taggerDao.addCollectionToCategory(id, category)
           .then(function () {
-            utils.sendResponse(res, {status: 'success'});
+            utils.sendSuccessJson(res);
 
           }).catch(function (err) {
           console.log(err);
@@ -548,7 +543,7 @@ exports.update = function (req, res) {
       } else {
         taggerDao.updateCollectionCategory(id, category).then(
           function () {
-            utils.sendResponse(res, {status: 'success'});
+            utils.sendSuccessJson(res);
 
           }).catch(
           function (err) {
@@ -649,7 +644,7 @@ exports.updateImage = function (req, res, config) {
 
     if (files.file !== undefined) {
       try {
-        imageConvert(res, files, fields, config, updateImageInDb)
+        imageConvert(res, files, fields, config, updateImageInDb);
       } catch (err) {
         console.log(err);
         res.end();
@@ -668,7 +663,7 @@ exports.updateImage = function (req, res, config) {
    */
   function updateImageInDb(res, id, imageName) {
     taggerDao.updateCollectionImage(id, imageName).then(function () {
-        utils.sendResponse(res, {status: 'success'});
+      utils.sendSuccessJson(res);
       }
     ).catch(function (err) {
         console.log(err);
@@ -676,237 +671,4 @@ exports.updateImage = function (req, res, config) {
     );
   }
 };
-
-// The following are rest endpoints for external clients (e.g. Academic Commons).
-
-/**
- * Retrieves the tags associated with a single collection. Used by
- * both admin interface and public REST API.
- * @param req
- * @param res
- */
-exports.tagsForCollection = function (req, res) {
-  var collId = req.params.collId;
-
-  taggerDao.findTagsForCollection(collId).then(function (tags) {
-    utils.sendResponse(res, tags);
-  }).catch(function (err) {
-    console.log(err);
-    utils.sendResponse(res, {status: 'failed'});
-  });
-
-};
-
-/**
- * Retrieves the types associated with a single collection.  Used by
- * both admin interface and public REST API.
- * @param req
- * @param res
- */
-exports.typesForCollection = function (req, res) {
-  var collId = req.params.collId;
-
-  taggerDao.findContentTypesForCollection(collId).then(function (types) {
-    utils.sendResponse(res, types);
-  })
-    .catch(function (err) {
-      console.log(err);
-      utils.sendResponse(res, {status: 'failed'});
-    });
-
-};
-
-/**
- * Retrieves a list of all collections for the public API.
- * @param req
- * @param res
- */
-exports.allCollections = function (req, res) {
-
-  taggerDao.retrieveAllCollections().then(function (collections) {
-    utils.sendResponse(res, collections)
-
-  }).catch(function (err) {
-    console.log(err);
-
-  });
-};
-
-/**
- * Retrieves single collection information for the public API.
- * @param req
- * @param res
- */
-exports.collectionById = function (req, res) {
-  const collId = req.params.id;
-
-  async.series({
-      collection: function (callback) {
-
-        taggerDao.findCollectionById(collId).then(
-          function (data) {
-            callback(null, data);
-          }).catch(
-          function (err) {
-            // trigger  error callback
-            callback(err);
-          });
-
-      },
-      categories: function (callback) {
-
-        taggerDao.getCategoryForCollection(collId).then(
-          function (data) {
-            callback(null, data);
-          }).catch(
-          function (err) {
-            callback(err);
-          });
-
-      },
-      itemTypes: function (callback) {
-
-        taggerDao.findContentTypesForCollection(collId).then(
-          function (data) {
-            callback(null, data);
-          }).catch(
-          function (err) {
-            callback(err);
-          });
-      }
-    },
-    function (err, result) {
-      if (err) {
-        console.log(err);
-        utils.sendResponse(res, {status: 'failed', reason: err});
-      } else {
-        utils.sendResponse(res, result);
-      }
-    }
-  );
-
-};
-
-/**
- * Retrieves list of collection by area ID for the public API.
- * @param req
- * @param res
- */
-exports.collectionsByArea = function (req, res) {
-  const areaId = req.params.id;
-
-  taggerDao.getCollectionsByArea(areaId).then(
-    function (collections) {
-      utils.sendResponse(res, collections);
-
-    }).catch(function (err) {
-    console.log(err);
-  });
-};
-
-/**
- * Retrieves a list of collections by subject and area for the public API.
- * @param req
- * @param res
- */
-exports.collectionsBySubject = function (req, res) {
-  const subjectId = req.params.id;
-  const areaId = req.params.areaId;
-
-  taggerDao.getCollectionsBySubjectAndArea(subjectId, areaId).then(
-    function (collections) {
-      utils.sendResponse(res, collections);
-    }).catch(function (err) {
-    console.log(err);
-  });
-
-};
-
-/**
- * Retrieves a list of collections by category for the public API.
- * @param req
- * @param res
- */
-exports.allCollectionsByCategory = function (req, res) {
-  const categoryId = req.params.id;
-
-  taggerDao.getCollectionsByCategory(categoryId).then(function (collections) {
-    utils.sendResponse(res, collections);
-  }).catch(function (err) {
-    console.log(err);
-  });
-
-};
-
-/**
- * Retrieves collections by subject (from all areas)
- */
-exports.allCollectionsBySubject = function (req, res) {
-  const subjectId = req.params.id;
-
-  taggerDao.getCollectionsBySubject(subjectId).then(
-    function (collections) {
-      utils.sendResponse(res, collections);
-    }).catch(function (err) {
-    console.log(err);
-  });
-};
-
-/**
- * Used by the Academic Commons.  Returns a JSON list of
- * objects retrieved from the eXist database host.  The fields
- * are the query term (title) and count.
- *
- * {
- *   item: {
- *     title: "1906",
- *     count: "4"
- * }
- *
- * Uses the 'collection' request parameter in the query.
- *
- * @param req
- * @param res
- */
-exports.browseList = function (req, res) {
-  const collection = req.params.collection;
-
-  const http = require('http');
-  const options = {
-    headers: {
-      accept: 'application/json'
-    },
-    host: config.externalHostA.host,
-    port: config.externalHostA.port,
-    path: config.externalHostA.path + collection,
-    method: 'GET'
-  };
-  // If no error, handle response.
-  function handleResponse(response) {
-
-    var str = '';
-    response.on('data', function (chunk) {
-      // Add data as it returns.
-      str += chunk;
-    });
-
-    response.on('end', function () {
-      res.setHeader('Content-Type', 'application/json');
-      res.setHeader('Access-Control-Allow-Origin', '*');
-      res.end(str);
-
-    });
-  }
-
-  const request = http.request(options, handleResponse);
-
-  request.on('error', function (e) {
-    console.log(e);
-    request.end();
-  });
-
-  request.end();
-};
-
-
