+++ server/api/tagger/dao/tag-target-dao.js
@@ -19,6 +19,8 @@
  * Created by mspalti on 12/1/16.
  */
 
+// jshint strict:false
+
 const taggerSchema = require('../models/index');
 const taggerDao = {};
 
@@ -90,7 +92,7 @@ taggerDao.removeTagFromArea = (areaId, tagId) => {
         AreaId: areaId
       }
     }
-  )
+  );
 
 };
