+++ server/api/tagger/utils/response-utility.js
@@ -19,12 +19,46 @@
  * Created by mspalti on 12/1/16.
  */
 
+//jshint strict:false
+
 const utils = {};
 
 utils.sendResponse = function(res, data) {
   res.setHeader('Content-Type', 'application/json');
   res.setHeader('Access-Control-Allow-Origin', '*');
   res.end(JSON.stringify(data));
+
+};
+
+/**
+ * Multiple components return area targets.  This utility provides
+ * a uniform interface.
+ * @param res
+ * @param err
+ * @param areas
+ */
+utils.sendAreaTargetsReponse = function (res, err, areas) {
+  if (err) {
+    console.log(err);
+  }
+  utils.sendResponse(res, {status: 'success', areaTargets: areas.areaList});
+};
+
+/**
+ * Provides uniform API for unadorned error responses.
+ * @param res
+ */
+utils.sendErrorJson = function (res, err) {
+  utils.sendResponse(res, {status: 'failed', reason: err.message});
+
+};
+
+/**
+ * Provides uniform API for unadorned success responses.
+ * @param res
+ */
+utils.sendSuccessJson = function (res) {
+  utils.sendResponse(res, {status: 'success'});
 
 };
